requirements.txt
requests
more-executors>=2.1.0
six
PyYAML
# For RefResolver.resolve
jsonschema>=2.5.0
attrs
six
